normalize test_account
